refactor(core): split analysis/candlePatterns.ts into subdirectory
Move CANDLE_PATTERN_META and detectCandlePatterns into
candlePatterns/meta.ts and candlePatterns/detect.ts, keeping
candlePatterns.ts as a re-export barrel so all existing import
sites (internal and @kansoku/core/analysis/candlePatterns) keep
working. No logic or detection order changes.
